requirements.txt
streamlit
plotly
pandas
google-api-python-client
google-auth
requests
ceic-api-client @ https://downloads.ceicdata.com/python/ceic-api-client/ceic_api_client-2.11.4.390.tar.gz
